Small refactoring of cache injections.
