+++ kifi-backend/modules/common/app/com/keepit/common/cache/FortyTwoCache.scala
@@ -308,4 +308,10 @@ class PrimitiveCacheImpl[K <: Key[P], P <: AnyVal](innermostPluginSettings: (For
 class StringCacheImpl[K <: Key[String]](innermostPluginSettings: (FortyTwoCachePlugin, Duration), innerToOuterPluginSettings: (FortyTwoCachePlugin, Duration)*)
   extends FortyTwoCacheImpl[K, String](innermostPluginSettings, innerToOuterPluginSettings:_*)(Serializer.string)
 
-trait CacheModule extends ScalaModule
+trait CachePluginModule extends ScalaModule
+abstract class CacheModule(cachePluginModules: CachePluginModule*) extends ScalaModule {
+  def configure() {
+    cachePluginModules.foreach(install)
+  }
+}
+
